Enabled parallel tasks multiprocessing

+++ config_verifications.py
@@ -20,86 +20,23 @@ from genie.testbed import load
 from genie.utils.diff import Diff
 from genie.utils import Dq
 from genie.utils.config import Config
-from unicon.core.errors import TimeoutError, StateMachineError, ConnectionError
+from common_setup_cleanup import CommonSetup as base_CommonSetup
+from common_setup_cleanup import CommonCleanup as base_CommonCleanup
 
 
 # create a logger for this module
 logger = logging.getLogger(__name__)
 
 
-class CommonSetup(aetest.CommonSetup):
+class CommonSetup(base_CommonSetup):
     # unwanted keys to delete from config
     keys_regex_to_delete_from_config = ( r'\S{3} \S{3} \d{1,2} \d{1,2}:\d{1,2}:\d{1,2}\.\d{1,3} \S{3}',
-                                r'\!\!\d{1,2}:\d{1,2}:\d{1,2} \S{3} \S{3} \S{3} \d{1,2} \d{4,}'
-                            )
+                                         r'\!\!\d{1,2}:\d{1,2}:\d{1,2} \S{3} \S{3} \S{3} \d{1,2} \d{4,}'
+                                        )
 
     @aetest.subsection
-    def verify_pre_directory(self, pre, post):
-        """
-        verify pre verifications report directory exist
-        """
-        if not post:
-            self.skipped("No need to check if pre verification report directory exist during pre verifications")
-        else:
-            if os.path.exists(pre):
-                self.passed(f"Pre verifications Report directory {pre} exists!")
-            else:
-                self.failed(f"Pre verifications Report directory {pre} does not exists!")
-
-
-    @aetest.subsection
-    def create_report_directory(self, pre, post):
-        """
-        create report directory
-        """
-
-        # where verification stage is pre or post
-        if pre and post:
-            report_directory = post
-        elif pre:
-            report_directory = pre
-        
-        # create report_directory if it does not exist
-        if os.path.exists(report_directory):
-            # To store the pre-existence status of the report directory
-            self.report_directory_alredy_exist = True
-            self.skipped(f"Report directory {report_directory} already exists.")
-        else:
-            # To store the pre-existence status of the report directory
-            self.report_directory_alredy_exist = False
-            try:
-                os.mkdir(report_directory)
-            except:
-                self.failed(f"There was an error while trying to create report directory {report_directory}")
-            self.passed(f"Report directory {report_directory} created successfully")
-
-
-    @aetest.subsection
-    def connect(self, testbed):
-        """
-        establishes connection to all your testbed devices.
-        """
-        # skipp the task if the report_directory_alredy_exist and the test case is in standalone execution
-        if self.report_directory_alredy_exist and __name__ == "__main__":
-            self.skipped("Don't need to connect if the report_directory_alredy_exist and the test case is in standalone execution.")
-
-        # make sure testbed is provided
-        assert testbed, "Testbed is not provided!"
-
-        # connect to all testbed devices
-        #   By default ANY error in the CommonSetup will fail the entire test run
-        #   Here we catch common exceptions if a device is unavailable to allow test to continue
-        try:
-            testbed.connect()
-        except (TimeoutError, StateMachineError, ConnectionError):
-            logger.error("Unable to connect to all devices")
-
-    @aetest.subsection
-    def learn_configs(self, testbed, steps, pre, post):
+    def learn_configs(self, testbed, steps, report_directory):
         """Learn and save configurations on the testbed devices."""
-        # skipp the task if the report_directory_alredy_exist and the test case is in in standalone execution
-        if self.report_directory_alredy_exist and __name__ == "__main__":
-            self.skipped("Don't need to learn config feature if the report_directory_alredy_exist and the test case is executed as standalone.")
 
         # Types of configs to learn
         config_types = { "config_running": "show running-config",
@@ -107,40 +44,54 @@ class CommonSetup(aetest.CommonSetup):
                     "config_failed_startup": "show configuration failed startup"
                     }
 
-        # where verification stage is pre or post
-        if pre and post:
-            report_directory = post
-        elif pre:
-            report_directory = pre
-
         # loop over every device in testbed
         for device_name, device in testbed.devices.items():
-            logger.info(f"{device_name} connected status: {device.connected}")
-            logger.info(f"Learning Configs for {device_name}")
-            # loop over every config type
-            for config_type in config_types:
-                # create new steps contects for each config type
-                with steps.start(
-                    f"Learning config_type: {config_type}  for device: {device_name}", continue_=True
-                ) as config_type_step:
-                    # Try  to learn configs on supported network operation systems
-                    try:
-                        # execute show config commad for config type and tokenize it
-                        config_str = device.execute(config_types[config_type])
-                        config_tree = Config(config_str)
-                        config_tree.tree()
-                        # delete unwanted keys
-                        for key_regex_to_delete in self.keys_regex_to_delete_from_config:
-                                t = tuple(Dq(config_tree.config).contains(key_regex_to_delete, regex=True).reconstruct())
-                                if t:
-                                    k = t[0]
-                                    del config_tree.config[k]
-                        # dump the config tree to json file
-                        with open(os.path.join(report_directory, f"{device_name}_{config_type}.json"), "w") as f:
-                            json.dump(config_tree.config, f ,sort_keys=True, indent=4)
-                    except:
-                        config_type_step.skipped(f"Coudn't learn config_type: {config_type}  for device: {device_name}.")
-                    config_type_step.passed(f"Successfully learned config_type: {config_type}  for device: {device_name}.")
+
+            # create new steps context for each device
+            info_message = f"Learning Configs for {device_name}."
+            with steps.start(info_message, continue_=True) as device_step:
+                logger.info(info_message)
+
+                # loop over every config type
+                for config_type in config_types:
+
+                    # output file name and path
+                    output_file_name = f"{device_name}_{config_type}.json"
+                    output_file_path = os.path.join(report_directory, output_file_name)
+
+                    # create new steps context for each config type
+                    info_message =f"Learning config_type: {config_type}  for device: {device_name}"
+                    with device_step.start(info_message, continue_=True) as config_type_step:
+                        logger.info(info_message)
+
+                        # skipp the task if the output files already exist
+                        if os.path.exists(output_file_path) and os.path.isfile(output_file_path):
+                            device_step.skipped(f"Output file: {output_file_path} already exist.")
+
+                        # fail the task if device is not connected
+                        elif not device.connected():
+                            device_step.failed(f"Device {device_name} is not connected.")
+
+                        else:
+                            # Try  to learn configs on supported network operation systems
+                            try:
+                                # execute show config commad for config type and tokenize it
+                                config_str = device.execute(config_types[config_type])
+                                config_tree = Config(config_str)
+                                config_tree.tree()
+                                # delete unwanted keys
+                                for key_regex_to_delete in self.keys_regex_to_delete_from_config:
+                                        t = tuple(Dq(config_tree.config).contains(key_regex_to_delete, regex=True).reconstruct())
+                                        if t:
+                                            k = t[0]
+                                            del config_tree.config[k]
+                                # dump the config tree to json file
+                                with open(output_file_path, "w") as f:
+                                    json.dump(config_tree.config, f ,sort_keys=True, indent=4)
+                            except BaseException as e:
+                                config_type_step.failed(f"Coudn't learn config_type: {config_type}  for device: {device_name}, reason: {e}.")
+                            config_type_step.passed(f"Successfully learned config_type: {config_type}  for device: {device_name}.")
+
 
 class compare_pre_post_config(aetest.Testcase):
     """compare_pre_post_config
@@ -153,14 +104,8 @@ class compare_pre_post_config(aetest.Testcase):
                     }
 
     @aetest.setup
-    def setup(self, testbed, steps, pre, post):
+    def setup(self, testbed, steps, pre, post, report_directory):
         """Read the learned configurations from json file for every the testbed devices."""
-
-        # where verification stage is pre or post
-        if pre and post:
-            report_directory = post
-        elif pre:
-            report_directory = pre
 
         # dicts to store device config information
         self.learnt_config = {}
@@ -174,13 +119,13 @@ class compare_pre_post_config(aetest.Testcase):
 
             # Trye to read pre learned config information
             with steps.start(
-                f"Reding pre saved configs from {device_name}", continue_=True
+                f"Reading pre saved configs from {device_name}", continue_=True
             ) as device_step:
 
                 # loop over every config type
                 for config_type in self.config_types:
                     with device_step.start(
-                        f"Reding config_type: {config_type}, from device: {device_name}", continue_=True
+                        f"Reading config_type: {config_type}, from device: {device_name}", continue_=True
                     ) as config_type_step:
                         try:
                             # read learned configs
@@ -242,24 +187,8 @@ class compare_pre_post_config(aetest.Testcase):
                             device_step.passed(f"No {failed_config_key} failed configurations found on device {device_name}")
 
 
-class CommonCleanup(aetest.CommonCleanup):
-    """CommonCleanup Section
-    < common cleanup docstring >
-    """
-
-    @aetest.subsection
-    def disconnect(self, testbed):
-        """
-        disconnects from all your testbed devices.
-        """
-        # make sure testbed is provided
-        assert testbed, "Testbed is not provided!"
-
-        # disconnect from all testbed devices
-        logger.info(
-            "Disconnecting from all testbed devices"
-        )
-        testbed.disconnect()
+class CommonCleanup(base_CommonCleanup):
+    pass
 
 
 if __name__ == "__main__":
@@ -295,8 +224,16 @@ if __name__ == "__main__":
         # type=Genie.init,
         default=None,
     )
+    parser.add_argument(
+        "--offline",
+        dest="offline",
+        help="offline mode",
+        type=bool,
+        # type=Genie.init,
+        default=None,
+    )
 
     # do the parsing
     args, sys.argv[1:] = parser.parse_known_args(sys.argv[1:])
 
-    aetest.main(testbed=args.testbed, pre=args.pre, post=args.post)
+    aetest.main(testbed=args.testbed, pre=args.pre, post=args.post, offline=args.offline)
